# services/internal/gen/http/openapi3.yaml
openapi: 3.0.3
info:
    title: Chatbook Microservices
    version: "1.0"
servers:
    - url: http://localhost:80
      description: Default server for chatbook-microservices
paths:
...
    /api/v1/users/profile:
        put:
            tags:
                - user
            summary: updateProfileNames user
            operationId: user#updateProfileNames
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateProfileNamesRequestBody'
                        example:
                            firstName: Quibusdam qui.
                            lastName: Tempora perspiciatis ut ut.
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/OperationStatusResponse'
                            example:
                                message: Optio a.
    /api/v1/users/register:
        post:
            tags:
                - user
            summary: register user
            operationId: user#register
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/RegisterRequestBody'
                        example:
                            confirmedPassword: Magnam facilis incidunt occaecati consequatur ullam.
                            email: Ducimus asperiores sunt.
                            firstName: Qui impedit tempore sunt optio.
                            lastName: Quisquam eos vitae velit quis.
                            password: Adipisci qui suscipit ut.
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/OperationStatusResponse'
                            example:
                                message: Quo et.
                "409":
                    description: 'UnmatchingPassowrds: Provided passwords are not the same'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
components:
    schemas:
        Error:
            type: object
            properties:
                fault:
                    type: boolean
                    description: Is the error a server-side fault?
                    example: true
                id:
                    type: string
                    description: ID is a unique identifier for this particular occurrence of the problem.
                    example: 123abc
                message:
                    type: string
                    description: Message is a human-readable explanation specific to this occurrence of the problem.
                    example: parameter 'p' must be an integer
                name:
                    type: string
                    description: Name is the name of this class of errors.
                    example: bad_request
                temporary:
                    type: boolean
                    description: Is the error temporary?
                    example: true
                timeout:
                    type: boolean
                    description: Is the error a timeout?
                    example: false
            example:
                id: 3F1FKVRR
                message: Value of ID must be an integer
                name: bad_request
            required:
                - name
                - id
                - message
                - temporary
                - timeout
                - fault
...
        OperationStatusResponse:
            type: object
            properties:
                message:
                    type: string
                    description: Operation status
                    example: Optio error est ut velit.
            example:
                message: Numquam molestiae dolorem.
            required:
                - message
...
        RegisterRequestBody:
            type: object
            properties:
                confirmedPassword:
                    type: string
                    description: Confirmed password of the user
                    example: Sit perferendis sapiente minima.
                email:
                    type: string
                    description: User email
                    example: Aliquid debitis amet officiis dicta ea.
                firstName:
                    type: string
                    description: User first name
                    example: Voluptate veniam incidunt inventore illum.
                lastName:
                    type: string
                    description: User last name
                    example: Delectus possimus nam accusamus.
                password:
                    type: string
                    description: User password
                    example: Molestiae omnis.
            example:
                confirmedPassword: Aliquid provident est illo officiis.
                email: Voluptatem amet fugit dolores sequi.
                firstName: Nulla sed.
                lastName: Minima error ad veniam pariatur qui.
                password: Dolorem dolorem et pariatur voluptas.
            required:
                - email
                - password
                - confirmedPassword
                - firstName
                - lastName
        UpdateProfileNamesRequestBody:
            type: object
            properties:
                firstName:
                    type: string
                    description: Updated user first name
                    example: Quia ab est repellendus exercitationem aut delectus.
                lastName:
                    type: string
                    description: Updated user last name
                    example: Ex repellendus.
            example:
                firstName: Sapiente nostrum minima reiciendis.
                lastName: Qui beatae deserunt nihil qui recusandae hic.
            required:
                - firstName
                - lastName
...